refactor Obsidian API helpers
