+++ src/components/CardViewer.tsx
@@ -2,8 +2,8 @@ import classcat from 'classcat';
 import { Component, MarkdownRenderer } from 'obsidian';
 import { useEffect, useRef } from 'react';
 import SRSCard from '#/lib/SRSCard';
-import { splitFrontMatter } from '#/lib/utils';
 import { useReviewContext } from './ReviewContext';
+import { ObsidianHelpers } from '#/lib/ObsidianHelpers';
 
 /** Read-only card viewer */
 export function CardViewer({ cardText }: { cardText: string }) {
@@ -24,7 +24,7 @@ export function CardViewer({ cardText }: { cardText: string }) {
   useEffect(() => {
     if (!containerRef.current) return;
 
-    const splitResult = splitFrontMatter(cardText);
+    const splitResult = ObsidianHelpers.splitFrontMatter(cardText);
     if (!splitResult) {
       throw new Error('Failed to parse frontmatter from note:\n' + cardText);
     }
